CollectionFilter cleanup
